# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Score Sculptor API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /credit-reports:
    get:
      operationId: getCreditReports
      tags: [credit-reports]
      summary: List credit reports
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CreditReport"
    post:
      operationId: createCreditReport
      tags: [credit-reports]
      summary: Create/upload credit report record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCreditReportBody"
      responses:
        "201":
          description: Report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditReport"
  /credit-reports/{id}:
    get:
      operationId: getCreditReport
      tags: [credit-reports]
      summary: Get credit report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditReport"

  # NOTIFICATIONS
...
  /ai/conversations:
    get:
      operationId: getAiConversations
      tags: [ai]
      summary: List AI conversations
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiConversation"
    post:
      operationId: createAiConversation
      tags: [ai]
      summary: Create new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAiConversationBody"
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiConversation"
  /ai/conversations/{id}/messages:
    get:
      operationId: getAiMessages
      tags: [ai]
      summary: Get messages for conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiMessage"
    post:
      operationId: sendAiMessage
      tags: [ai]
      summary: Send a message to Sculpt AI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendAiMessageBody"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiMessage"

  # EDUCATION
...
components:
  schemas:
...
    RegisterBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
      required: [email, password, firstName, lastName]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [user, admin]
        creditGoal:
          type: string
          nullable: true
        educationLevel:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, email, firstName, lastName, role, createdAt]
...
    CreditReport:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        fileName:
          type: string
        bureau:
          type: string
          nullable: true
        reportDate:
          type: string
          nullable: true
        creditScore:
          type: integer
          nullable: true
        status:
          type: string
          enum: [uploaded, analyzing, analyzed, error]
        analysisNotes:
          type: string
          nullable: true
        inconsistenciesCount:
          type: integer
          nullable: true
        createdAt:
          type: string
      required: [id, userId, fileName, status, createdAt]
    CreateCreditReportBody:
      type: object
      properties:
        fileName:
          type: string
        bureau:
          type: string
        reportDate:
          type: string
        creditScore:
          type: integer
      required: [fileName]

    # NOTIFICATION SCHEMAS
...
    AiConversation:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, title, createdAt, updatedAt]
    CreateAiConversationBody:
      type: object
      properties:
        title:
          type: string
      required: [title]
    AiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
      required: [id, conversationId, role, content, createdAt]
    SendAiMessageBody:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # EDUCATION SCHEMAS
...