# lib/api-zod/src/generated/types/createAiConversationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Score Sculptor API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAiConversationBody {
  title: string;
}